requirements.txt
jupyter==1.1.1
deep-translator==1.11.4
pandas==2.2.3
numpy==2.0.2
requests==2.32.3
scikit-learn==1.5.2
keras==3.7.0
tensorflow==2.18.0
Flask-admin==1.6.1
Flask==3.1.0
python-dotenv==1.0.1
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1